% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{ca}
\alias{ca}
\title{Correspondence analysis}
\usage{
ca(DATA, k = 0, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{DATA}{a data matrix to decompose}

\item{k}{total number of components to return (see \code{\link{gsvd}}).}

\item{tol}{default is .Machine$double.eps. A tolerance level for eliminating effectively zero (small variance), negative, imaginary eigen/singular value components (see \code{\link{gsvd}}).}
}
\value{
A list with eleven elements by way of \code{\link[GSVD]{gsvd}}:
\item{d.orig}{A vector containing the singular values of X above the tolerance threshold (based on eigenvalues).}
\item{l.orig}{A vector containing the eigen values of X above the tolerance threshold (\code{tol}).}
\item{tau}{A vector that contains the (original) explained variance per component (via eigenvalues: \code{$l.orig}).}
\item{d}{A vector of length \code{min(length(d.orig), k)} containing the retained singular values of X}
\item{l}{A vector of length \code{min(length(l.orig), k)} containing the retained eigen values of X}
\item{u}{Left (rows) singular vectors. Dimensions are \code{nrow(DATA)} by k.}
\item{p}{Left (rows) generalized singular vectors. Dimensions are \code{nrow(DATA)} by k.}
\item{fi}{Left (rows) component scores. Dimensions are \code{nrow(DATA)} by k.}
\item{v}{Right (columns) singular vectors. Dimensions are \code{ncol(DATA)} by k.}
\item{q}{Right (columns) generalized singular vectors. Dimensions are \code{ncol(DATA)} by k.}
\item{fj}{Right (columns) component scores. Dimensions are \code{ncol(DATA)} by k.}
}
\description{
\code{ca} performs correspondence analysis of a data matrix \code{DATA}.
}
\examples{


}
\seealso{
\code{\link{gsvd}}
}
\author{
Derek Beaton
}
