% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{ca_preproc}
\alias{ca_preproc}
\title{Correspondence analysis (CA) preprocessing}
\usage{
ca_preproc(DATA)
}
\arguments{
\item{DATA}{a numeric matrix}
}
\value{
a list with 6 elements
\item{Ox:} {The observed values}
\item{m:} {Row probabilities}
\item{w:} {Column probabilities}
\item{Ex:} {The expected values}
\item{Zx:} {The deviations values}
\item{weightedZx:} {The deviations divided by the square root of the row and column probabilities}
}
\description{
Preprocessing/preparing a data matrix for correspondence analysis
}
\details{
A data matrix processed akin to the way \eqn{\chi^2} is computed (deviations from independence)
}
\author{
Derek Beaton
}
