% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmaxs.R
\name{categorical_corrmax}
\alias{categorical_corrmax}
\title{CorrMax transformation for categorical data}
\usage{
categorical_corrmax(target.data, loadings, singular.values)
}
\arguments{
\item{target.data}{a}

\item{loadings}{a}

\item{singular.values}{a}
}
\description{
A transformation of the data in a row-wise fashion that preserves the Mahalanobis distances yet also provides information on which variables likely contribute to the Mahalanobis distance of the observation
}
\author{
Derek Beaton
}
