% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{component_plot}
\alias{component_plot}
\title{Plot components from PCA or CA}
\usage{
component_plot(scores, axes = c(1, 2), pch = 20,
  col = "mediumorchid4", main = "Component scores",
  xlab = paste0("Component ", axes[1]), ylab = paste0("Component ",
  axes[2]), xlim = c(-max(abs(scores[, axes])), max(abs(scores[, axes])))
  * 1.3, ylim = c(-max(abs(scores[, axes])), max(abs(scores[, axes]))) *
  1.3, asp = 1, pos = 3, display_names = T, cex = 1,
  text.cex = 1, ...)
}
\arguments{
\item{scores}{a numeric matrix with the scores (fi or fj) or vectors (u/p or v/q)}

\item{axes}{a numeric vector of length 2. Indicates which axes (columns) of \code{scores} to plot}

\item{pch}{see \code{\link{plot}} and \code{\link{par}}. Default here is '20'}

\item{col}{see \code{\link{plot}} and \code{\link{par}}. Default here is 'mediumorchid4'}

\item{main}{see \code{\link{plot}} and \code{\link{par}}. Default here is 'Component scores'}

\item{xlab}{see \code{\link{plot}} and \code{\link{par}}. Default here is '\code{paste0("Component ",axes[1])}'}

\item{ylab}{see \code{\link{plot}} and \code{\link{par}}. Default here is '\code{paste0("Component ",axes[2])}'}

\item{xlim}{see \code{\link{plot}} and \code{\link{par}}. Default here is '\code{c(-max(abs(scores[,axes])),max(abs(scores[,axes])))*1.3}'}

\item{ylim}{see \code{\link{plot}} and \code{\link{par}}. Default here is '\code{c(-max(abs(scores[,axes])),max(abs(scores[,axes])))*1.3}'}

\item{asp}{see \code{\link{plot}} and \code{\link{par}}. Default here is '1'}

\item{pos}{see \code{\link{plot}} and \code{\link{par}}. Default here is '3'}

\item{display_names}{a logical (boolean). If \code{TRUE} plot the row names of \code{scores}}

\item{cex}{see \code{\link{plot}} and \code{\link{par}}. Default here is '1'}

\item{text.cex}{like \code{cex} but for text plotting when \code{display_names = TRUE}. See also \code{\link{plot}} and \code{\link{par}}. Default here is '1'}

\item{...}{Additional arguments to be passed for plotting. See \code{\link{plot}} and \code{\link{par}}.}
}
\description{
a simple 2D plot for observation (or variable) component scores or vectors to visualize PCA or CA
}
\author{
Derek Beaton
}
