% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmaxs.R
\name{continous_corrmax}
\alias{continous_corrmax}
\title{CorrMax transformation for continuous data}
\usage{
continous_corrmax(target.data, center = T, scale = F, loadings,
  singular.values)
}
\arguments{
\item{target.data}{a}

\item{center}{a}

\item{scale}{a}

\item{loadings}{a}

\item{singular.values}{a}
}
\description{
A transformation of the data in a row-wise fashion that preserves the Mahalanobis distances yet also provides information on which variables likely contribute to the Mahalanobis distance of the observation
}
\examples{

}
\author{
Derek Beaton
}
