% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_mcd.R
\name{continuous_mcd}
\alias{continuous_mcd}
\title{Minimum covariance determinant (MCD) for continuous data}
\usage{
continuous_mcd(DATA, center = T, scale = F, allow_collinearity = F,
  alpha = 0.75, num.subsets = 500, max.total.iters = num.subsets *
  20, top.sets.percent = 0.05, tol = .Machine$double.eps)
}
\arguments{
\item{DATA}{a data matrix to decompose}

\item{center}{logical or numeric (see \code{\link{scale}}). Default is \code{TRUE} which centers the columns (e.g., when \code{TRUE} substract the mean of a column from its respective column)}

\item{scale}{logical or numeric (see \code{\link{scale}}). Default is \code{TRUE} which scales the columns (e.g., when \code{TRUE} divide a column by its respective standard deviation or scaling factor)}

\item{tol}{default is .Machine$double.eps. A tolerance level for eliminating effectively zero (small variance), negative, imaginary eigen/singular value components (see \code{\link{gsvd}}).}
}
\value{
A list of three lists:
\item \strong{cov:} a list for the robust covariance structure items
\itemize{
  \item{loadings} {a matrix of loadings via PCA from the robust covariance matrix}
  \item{singular.values} {a vector of singular values via PCA from the robust covariance matrix}
  \item{center} {a vector of the column-wise centers for the final subsample that produces the robust covariance matrix}
  \item{scale} {a vector of the column-wise scale for the final subsample that produces the robust covariance matrix}
}
\item{dists}{A vector containing the eigen values of X above the tolerance threshold (\code{tol}).}
\item{det.samps}{A vector that contains the (original) explained variance per component (via eigenvalues: \code{$l.orig}).}
}
\description{
\code{continuous_mcd} performs the MCD of a data matrix \code{DATA}.
}
\examples{


}
\seealso{
\code{\link{continuous_mcd_find_sample}}, \code{\link{categorical_mcd}}, \code{\link{ordinal_mcd}}, \code{\link{mixed_data_mcd}}, \code{\link{generalized_mcd}}
}
\author{
Derek Beaton
}
