% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_mcd.R
\name{continuous_mcd_find_sample}
\alias{continuous_mcd_find_sample}
\title{@export}
\usage{
continuous_mcd_find_sample(DATA, center = T, scale = F, alpha = 0.75,
  num.subsets = 500, max.total.iters = num.subsets * 20,
  top.sets.percent = 0.05, tol = .Machine$double.eps)
}
\description{
@export
}
