% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{dd_plot}
\alias{dd_plot}
\title{Distance-distance plot}
\usage{
dd_plot(ours_mcd_list, md_cutoff = NA, robust_md_cutoff = NA,
  dist_transform = "none")
}
\arguments{
\item{ours_mcd_list}{an OuRS MCD class object (of type list). Returned from \code{continuous_mcd}, \code{categorical_mcd}, \code{ordinal_mcd}, or \code{mixed_data_mcd}}

\item{md_cutoff}{numeric. A value for Mahalanobis distances to display a cutoff. If invalid, \code{quantile(x, probs=.95)} used}

\item{robust_md_cutoff}{numeric. A value for robust Mahalanobis distances to display a cutoff. If invalid, \code{quantile(x, probs=.95)} used}

\item{dist_transform}{character of type "none", "sqrt", or "log". If "sqrt" the distances will be transformed by square root, if "log" the distances will be transformed by the natural log. If "none" , no transformation will be performed.}
}
\description{
A plot of standard vs. robust Mahalanobis distances (with optional transformations)
}
\author{
Derek Beaton
}
