% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{disjunctive_coding}
\alias{disjunctive_coding}
\title{Complete disjunctive coding}
\usage{
disjunctive_coding(DATA, impute_NA_to_mean = F)
}
\arguments{
\item{DATA}{a matrix where each column is a variable, and each value within the columns is assumed to be a level within that variable}

\item{impute_NA_to_mean}{a logical (boolean). Default is \code{FALSE}. If \code{TRUE} all \code{NA} will be replaced with the column-wise mean *after* transformation.}
}
\description{
Complete disjunctive coding for categorical (nominal) data
}
\details{
If \code{impute_NA_to_mean = TRUE} then all \code{NA} will be replaced with the column-wise mean *after* transformation
The presence of a level is indicated with a '1', where the absence of a level is '0'. Imputed values are the mean proportion of 0s & 1s and thus (0,1).
}
\references{

}
\seealso{
\code{\link{escofier_coding}} and \code{\link{thermometer_coding}}
}
\author{
Derek Beaton
}
