% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{escofier_coding}
\alias{escofier_coding}
\title{Escofier's approach to "data-doubling" for continuous data}
\usage{
escofier_coding(DATA, center = TRUE, scale = TRUE, impute_NA_to_mean = F)
}
\arguments{
\item{DATA}{a numeric matrix}

\item{center}{logical or numeric (see \code{\link{scale}}). Default is \code{TRUE} which centers the columns (e.g., when \code{TRUE} substract the mean of a column from its respective column)}

\item{scale}{logical or numeric (see \code{\link{scale}}). Default is \code{TRUE} which scales the columns (e.g., when \code{TRUE} divide a column by its respective standard deviation or scaling factor)}

\item{impute_NA_to_mean}{a logical (boolean). Default is \code{FALSE}. If \code{TRUE} all \code{NA} will be replaced with the column-wise mean *after* centering and scaling.}
}
\description{
Performs data-doubling for continuous data so that the data behave like disjunctive data.
This function makes use of \code{\link{ours_scale}} to center and/or scale (presumably) continous data.
}
\details{
Data are returned in a "doubled" way where each column is represented twice as \eqn{[\frac{1-x}{2}, \frac{1+x}{2}]} where \code{x} is a centered and/or scaled vector.
If \code{impute_NA_to_mean = TRUE} then all \code{NA} will be replaced with the column-wise mean *after* centering/scaling
}
\references{

}
\seealso{
\code{\link{scale}}, \code{\link{disjunctive_coding}}, and \code{\link{thermometer_coding}}
}
\author{
Derek Beaton
}
