% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmaxs.R
\name{generalized_corrmax}
\alias{generalized_corrmax}
\title{CorrMax transformation for correspondence analysis-like data}
\usage{
generalized_corrmax(target.data, loadings, singular.values)
}
\arguments{
\item{target.data}{a}

\item{loadings}{a}

\item{singular.values}{a}
}
\description{
A transformation of the data in a row-wise fashion that preserves the Mahalanobis distances yet also provides information on which variables likely contribute to the Mahalanobis distance of the observation
}
\details{
\code{target.data} are assumed to be correspondence analysis (CA)-like data (i.e., counts). This function performs CA preprocessing before applying the CorrMax procedure.
This particular function is the primary function for \code{\link{categorical_corrmax}}, \code{\link{ordinal_corrmax}}, and \code{\link{mixed_data_corrmax}}.
}
\author{
Derek Beaton
}
