% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_mcd.R
\name{generalized_scores_dists}
\alias{generalized_scores_dists}
\title{Compute scores and distances with projection for generalized MCD}
\usage{
generalized_scores_dists(
  DATA,
  row.weights,
  col.weights,
  loadings,
  singular.values
)
}
\arguments{
\item{DATA}{a data matrix (of presumably all continuous data)}

\item{row.weights}{a numeric vector of row weights (as in correspondence analysis)}

\item{col.weights}{a numeric vector of column weights (as in correspondence analysis)}

\item{loadings}{a numeric matrix that contains the loadings (singular vectors or eigenvectors of variables)  from a decomposed covariance matrix}

\item{singular.values}{a numeric vector that contains the singular values from a decomposed covariance matrix}
}
\value{
a list with four items. All items are for the rows of \code{DATA} and computed through projection (via \code{loadings} and \code{singular.vectors})
\item{projected_u:} {Projected singular vectors}
\item{projected_fi:} {Projected component scores}
\item{projected_mahal_dists:} {Mahalanobis distances (computed as \code{rowSums(projected_u^2)})}
\item{projected_score_dists:} {Score distances (computed as \code{rowSums(projected_fi^2)})}
}
\description{
Computes projected: singular vectors, Mahalanobis distance, component scores, and score distances.
}
\details{
This approach uses loadings (singular vectors) and singular values to compute scores and distances for \code{DATA}. For use with the GMCD, \code{loadings} and \code{singular.values} come from a robust covariance matrix, where \code{row.weights} and \code{col.weights} are the weights from preprocessing for Correspondence Analysis.
}
\seealso{
\code{\link{generalized_mcd}} and \code{\link{continuous_scores_dists}}
}
\author{
Derek Beaton
}
