% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{mixed_data_coding}
\alias{mixed_data_coding}
\title{Mixed data coding}
\usage{
mixed_data_coding(
  DATA,
  column.type = rep("x", ncol(DATA)),
  impute_NA_to_mean = F
)
}
\arguments{
\item{DATA}{a matrix where each column is a variable, and each value within the columns is assumed to be a level within that variable}

\item{column.type}{a character vector of length \code{ncol(DATA)}. Each element indicates the data type of its respective column: "n" is categorical/nominal, "c" is continuous (only center the data), "z" is continuous (center and scale the data) "o" is ordinal, and "x" means 'do nothing to this column'. Each column will be transformed according to \code{disjunctive_coding} (categorical), \code{escofier_coding} (continuous), \code{thermometer_coding} (ordinal), or no transformation.}

\item{impute_NA_to_mean}{a logical (boolean). Default is \code{FALSE}. If \code{TRUE} all \code{NA} will be replaced with the column-wise mean *after* transformation.}
}
\description{
Transforms mixtures of data contained within a data.frame
}
\details{
For each type of data, this function calls into the respective transformation function.
\itemize{
\item "n" - is categorical data and calls \code{disjunctive_coding}
\item "c" - is continuous data that should only be centered and calls \code{escofier_coding} with \code{center = TRUE} and \code{scale = FALSE}
\item "z" - is continuous data that should be centered and scaled, and calls \code{escofier_coding} with \code{center = TRUE} and \code{scale = TRUE}
\item "o" - is ordinal data and calls \code{thermometer_coding}
\item "x" - is 'do nothing' and returns the data exactly as they are
}
The output (transformed) data are returned in a different order than they are input. The order matches the types as: "n", "c", "z", "o", and then "x"
In all cases, values are derived from the data and other parameters of those functions are not available. For examples: you cannot pass \code{min} to \code{thermometer_coding} here. Instead, the observed minimums will be used
\code{impute_NA_to_mean} is applied globally and passed into each function exactly as it is used here.
If \code{impute_NA_to_mean = TRUE} then all \code{NA} will be replaced with the column-wise mean *after* transformation
}
\seealso{
\code{\link{disjunctive_coding}}, \code{\link{thermometer_coding}}, and \code{\link{escofier_coding}}
}
\author{
Derek Beaton
}
