% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmaxs.R
\name{mixed_data_corrmax}
\alias{mixed_data_corrmax}
\title{CorrMax transformation for mixed data}
\usage{
mixed_data_corrmax(target.data, column.types = rep("x",
  ncol(target.data)), loadings, singular.values)
}
\arguments{
\item{target.data}{a}

\item{column.types}{a}

\item{loadings}{a}

\item{singular.values}{a}
}
\description{
A transformation of the data in a row-wise fashion that preserves the Mahalanobis distances yet also provides information on which variables likely contribute to the Mahalanobis distance of the observation
}
\examples{

}
\author{
Derek Beaton
}
