% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OuRS.R
\docType{package}
\name{ours-package}
\alias{ours-package}
\alias{_PACKAGE}
\alias{ours}
\alias{OuRS}
\alias{OURS}
\title{OuRS}
\description{
Outliers and Robust Structures
}
\details{
Some level of detail!
}
\references{

}
\seealso{
\code{\link{gsvd}}
}
\author{
\strong{Maintainer}: Derek Beaton \email{exposition.software@gmail.com}

Authors:
\itemize{
  \item Kelly Sunderland \email{ksunderland@research.baycrest.org}
}

}
