% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{ours_scale}
\alias{ours_scale}
\title{Variant of \code{scale}: scaling and centering of matrix-like objects}
\usage{
ours_scale(DATA, center = TRUE, scale = TRUE, impute_NA_to_mean = F)
}
\arguments{
\item{DATA}{a numeric matrix}

\item{center}{logical or numeric (see \code{\link{scale}}). Default is \code{TRUE} which centers the columns (e.g., when \code{TRUE} substract the mean of a column from its respective column)}

\item{scale}{logical or numeric (see \code{\link{scale}}). Default is \code{TRUE} which scales the columns (e.g., when \code{TRUE} divide a column by its respective standard deviation or scaling factor)}

\item{impute_NA_to_mean}{a logical (boolean). Default is \code{FALSE}. If \code{TRUE} all \code{NA} will be replaced with the column-wise mean *after* centering and scaling.}
}
\description{
Performs the same functionality as \code{\link{scale}} but always returns the "scale:center" and "scale:scale" attributes. Also allows mean imputation for \code{NA}
}
\details{
If no centering is performed (i.e., set to \code{FALSE}), then \code{attributes()`scale:center` <- rep(0, ncol(DATA))}
If no scaling is performed (i.e., set to \code{FALSE}), then \code{attributes()`scale:scale` <- rep(1, ncol(DATA))}
If \code{impute_NA_to_mean = TRUE} then all \code{NA} will be replaced with the column-wise mean *after* centering/scaling
}
\seealso{
\code{\link{scale}}
}
\author{
Derek Beaton
}
