% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{thermometer_coding}
\alias{thermometer_coding}
\title{"Thermometer" approach to "data-doubling" for ordinal data}
\usage{
thermometer_coding(DATA, mins, maxs, impute_NA_to_mean = F)
}
\arguments{
\item{DATA}{a numeric matrix}

\item{mins}{(optional) numeric vector of length \code{ncol(DATA)}. Contains the expected minimum value per column. If none provided then they are computed as \code{apply(DATA, 2, min, na.rm = T)}}

\item{maxs}{(optional) numeric vector of length \code{ncol(DATA)}. Contains the expected maximum value per column. If none provided then they are computed as \code{apply(DATA, 2, max, na.rm = T)}}

\item{impute_NA_to_mean}{a logical (boolean). Default is \code{FALSE}. If \code{TRUE} all \code{NA} will be replaced with the column-wise mean *after* transformation.}
}
\description{
Performs data-doubling for ordinal data so that the data behave like disjunctive data.
}
\details{
Data are returned in a "doubled" way where each column is represented twice as \eqn{[\frac{1-x}{2}, \frac{1+x}{2}]} where \code{x} is a centered and/or scaled vector.
If \code{impute_NA_to_mean = TRUE} then all \code{NA} will be replaced with the column-wise mean *after* transformation.
}
\references{

}
\seealso{
\code{\link{escofier_coding}} and \code{\link{disjunctive_coding}}
}
\author{
Derek Beaton
}
