% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{tolerance_ellipse}
\alias{tolerance_ellipse}
\title{Make tolerance ellipse for a 2D or x-y scatter plot}
\usage{
tolerance_ellipse(DATA, ellipse.alpha = 0.75, mcd.alpha = 0.75,
  xlab = colnames(DATA)[1], ylab = colnames(DATA)[2], graphs = F)
}
\arguments{
\item{DATA}{a numeric matrix with (presumably) two columns. This function will only make use of the first two.}

\item{ellipse.alpha}{numeric in the range of (.5,1). The percentage for a tolerance ellipse}

\item{mcd.alpha}{numeric in the range of (.5,1).  The percentage for a robust tolerance ellipse (by way of the MCD)}

\item{xlab}{see \code{\link{plot}} and \code{\link{par}}. Default here is '\code{colnames(DATA)[1]}'}

\item{ylab}{see \code{\link{plot}} and \code{\link{par}}. Default here is '\code{colnames(DATA)[2]}'}

\item{graphs}{logical (boolean). Default is \code{FALSE}. When \code{FALSE} ellipses are added to an existing plot, when \code{TRUE} a new x-y plot is made with ellipses}
}
\description{
Compute the tolerance ellipse for x-y data
}
\details{
Note: This function is effectively a wrapper for a set of functions copied from the R package SIBER.
Here, many of the SIBER functions are private to \code{tolerance_ellipse}, but were copied directly from SIBER 2.1.0. 
So users of \code{OuRS} only interface with \code{tolerance_ellipse} and not any of the private functions
}
\seealso{
https://CRAN.R-project.org/package=SIBER
}
\author{
Derek Beaton for \code{tolerance_ellipse}; Andrew Jackson and Andrew Parnell for all functions inside (i.e., \code{pointsToEllipsoid}, \code{ellipsoidTransform}, \code{ellipseInOut}, \code{addEllipse}, \code{genCircle})
}
