% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{ca_preproc}
\alias{ca_preproc}
\title{Correspondence analysis (CA) preprocessing}
\usage{
ca_preproc(DATA, compact = T)
}
\arguments{
\item{DATA}{a numeric matrix}

\item{compact}{a logical (boolean). Default is \code{TRUE}. When \code{TRUE} only some elements are returned}
}
\value{
\item if \code{compact = FALSE} a list with 6 elements
\itemize{
  \item{Ox:} {The observed values}
  \item{m:} {Row probabilities}
  \item{w:} {Column probabilities}
  \item{Ex:} {The expected values}
  \item{Zx:} {The deviations values}
  \item{weightedZx:} {The deviations divided by the square root of the row and column probabilities}
}
\item if \code{compact = TRUE} a list with 3 elements
\itemize{
  \item{m:} {Row probabilities}
  \item{w:} {Column probabilities}
  \item{weightedZx:} {The deviations divided by the square root of the row and column probabilities}
}
}
\description{
Preprocessing/preparing a data matrix for correspondence analysis
}
\details{
A data matrix processed akin to the way \eqn{\chi^2} is computed (deviations from independence)
}
\author{
Derek Beaton
}
