% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_mcd.R
\name{continuous_scores_dists}
\alias{continuous_scores_dists}
\title{Compute scores and distances with projection for continuous MCD}
\usage{
continuous_scores_dists(DATA, center = T, scale = F, loadings, singular.values)
}
\arguments{
\item{DATA}{a data matrix (of presumably all continuous data)}

\item{center}{logical or numeric (see \code{\link{scale}}). Default is \code{TRUE} which centers the columns (e.g., when \code{TRUE} substract the mean of a column from its respective column)}

\item{scale}{logical or numeric (see \code{\link{scale}}). Default is \code{TRUE} which scales the columns (e.g., when \code{TRUE} divide a column by its respective standard deviation or scaling factor)}

\item{loadings}{a numeric matrix that contains the loadings (singular vectors or eigenvectors of variables)  from a decomposed covariance matrix}

\item{singular.values}{a numeric vector that contains the singular values from a decomposed covariance matrix}
}
\value{
a list with four items. All items are for the rows of \code{DATA} and computed through projection (via \code{loadings} and \code{singular.vectors})
\itemize{
\item projected_u - Projected singular vectors
\item projected_fi - Projected component scores
\item projected_mahal_dists - Mahalanobis distances (computed as \code{rowSums(projected_u^2)})
\item projected_score_dists - Score distances (computed as \code{rowSums(projected_fi^2)})
}
}
\description{
Computes projected: singular vectors, Mahalanobis distance, component scores, and score distances.
}
\details{
This approach uses loadings (singular vectors) and singular values to compute scores and distances for \code{DATA}. For use with the MCD, \code{loadings} and \code{singular.values} come from a robust covariance matrix.
}
\seealso{
\code{\link{continuous_mcd}} and \code{\link{generalized_scores_dists}}
}
\author{
Derek Beaton
}
